+++ Data_Structures/C++/merge_sort.cpp
@@ -1,61 +1,96 @@
-// merge sort
-# include <iostream>
-using namespace std;
-
-void merge(int A[], int low, int q, int high) { 
-    int i, j, k; 
-    int n1 = q - low + 1; 
-    int n2 =  high - q; 
-    int L[n1], R[n2]; 
-    for (i = 0; i < n1; i++) 
-        L[i] = A[low + i]; 
-    for (j = 0; j < n2; j++) 
-        R[j] = A[q + 1+ j]; 
-    i = 0; 
-    j = 0; 
-    k = low; 
-    while (i < n1 && j < n2) 
-    { 
-        if (L[i] <= R[j]) 
-        { 
-            A[k++] = L[i++]; 
-        } 
-        else
-        { 
-            A[k++] = R[j++]; 
-        } 
-    } 
-    while (i < n1) 
-    { 
-        A[k++] = L[i++]; 
-    } 
-    while (j < n2) 
-    { 
-        A[k++] = R[j++];  
-    } 
-}
-
-void mergeSort(int A[], int low, int high) 
-{ 
-  if (low < high) 
-  { 
-      int q = low+(high-low)/2; 
-      mergeSort(A, low, q); 
-      mergeSort(A, q+1, high); 
-      merge(A, low, q, high); 
-      } 
+#include <stdio.h> 
+#include <stdlib.h> 
   
-}  
-
-int main()  
-{  
-    int A[] ={67, 23, 65, 10, 9,-12, 0};
-    int n = sizeof(A) / sizeof(A[0]);
-    mergeSort(A,0, n);  
-    cout<<"Sorted Array in increasing order is : \n"; 
-    int i;  
-    for (i = 0; i < n; i++)  
-        cout << A[i] << " ";  
-    cout << endl; 
-    return 0;  
-}  
+// Merges two subarrays of arr[]. 
+// First subarray is arr[l..m] 
+// Second subarray is arr[m+1..r] 
+void merge(int arr[], int l, int m, int r) 
+{ 
+    int i, j, k; 
+    int n1 = m - l + 1; 
+    int n2 = r - m; 
+  
+    /* create temp arrays */
+    int L[n1], R[n2]; 
+  
+    /* Copy data to temp arrays L[] and R[] */
+    for (i = 0; i < n1; i++) 
+        L[i] = arr[l + i]; 
+    for (j = 0; j < n2; j++) 
+        R[j] = arr[m + 1 + j]; 
+  
+    /* Merge the temp arrays back into arr[l..r]*/
+    i = 0; // Initial index of first subarray 
+    j = 0; // Initial index of second subarray 
+    k = l; // Initial index of merged subarray 
+    while (i < n1 && j < n2) { 
+        if (L[i] <= R[j]) { 
+            arr[k] = L[i]; 
+            i++; 
+        } 
+        else { 
+            arr[k] = R[j]; 
+            j++; 
+        } 
+        k++; 
+    } 
+  
+    /* Copy the remaining elements of L[], if there 
+       are any */
+    while (i < n1) { 
+        arr[k] = L[i]; 
+        i++; 
+        k++; 
+    } 
+  
+    /* Copy the remaining elements of R[], if there 
+       are any */
+    while (j < n2) { 
+        arr[k] = R[j]; 
+        j++; 
+        k++; 
+    } 
+} 
+  
+/* l is for left index and r is right index of the 
+   sub-array of arr to be sorted */
+void mergeSort(int arr[], int l, int r) 
+{ 
+    if (l < r) { 
+        // Same as (l+r)/2, but avoids overflow for 
+        // large l and h 
+        int m = l + (r - l) / 2; 
+  
+        // Sort first and second halves 
+        mergeSort(arr, l, m); 
+        mergeSort(arr, m + 1, r); 
+  
+        merge(arr, l, m, r); 
+    } 
+} 
+  
+/* UTILITY FUNCTIONS */
+/* Function to print an array */
+void printArray(int A[], int size) 
+{ 
+    int i; 
+    for (i = 0; i < size; i++) 
+        printf("%d ", A[i]); 
+    printf("\n"); 
+} 
+  
+/* Driver program to test above functions */
+int main() 
+{ 
+    int arr[] = { 12, 11, 13, 5, 6, 7 }; 
+    int arr_size = sizeof(arr) / sizeof(arr[0]); 
+  
+    printf("Given array is \n"); 
+    printArray(arr, arr_size); 
+  
+    mergeSort(arr, 0, arr_size - 1); 
+  
+    printf("\nSorted array is \n"); 
+    printArray(arr, arr_size); 
+    return 0; 
+}
