+++ Data_Structures/C/linkedlist.c
